Implemented deactivation and reactivation of users (replaced removal)
Implemented deactivation and reactivation of users by adding the deactivated_at field to the User model. Deactivated users can no longer log in and do no longer show up in the autocomplete list when composing a Kudo transaction. Updated the administrator dashboard to support this functionality (HAML code is automatically generated and customized).