Updated the functionality and appearance of the administrator dashboard
Updated the administrate gem which improves the design of the administrator dashboard. Added the administrate-field-image gem to show profile pictures of users on their profile pages. Added the Votes page to the administrator dashboard which lets an administrator view and remove votes. Regenerated and edited the admin/users views to support deactivation and reactivation of users.